refactor(state): move character shapes to characterStore.types.ts
characterStore.ts and lib/utils/derivedStatCalculator.ts imported each other
(the store needs calculateDerivedStat, the calculator needed the
CharacterAttribute type). Moving the plain data shapes into their own module
breaks that cycle; the store re-exports them so existing import sites keep
working.
