+++ src/lib/utils/derivedStatCalculator.ts
@@ -15,7 +15,7 @@
  */
 
 import { DerivedStatFormula } from '@/types/world.types';
-import { CharacterAttribute } from '@/state/characterStore';
+import { CharacterAttribute } from '@/state/characterStore.types';
 
 /**
  * Calculate a derived stat value from a formula and character attributes
