+++ src/state/characterStore.ts
@@ -2,7 +2,6 @@ import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
 import type { UseBoundStore, StoreApi } from 'zustand';
 import { EntityID } from '../types/common.types';
-import { InventoryItem, InventoryCategory } from '../types/inventory.types';
 import { DerivedStat } from '../types/character.types';
 import { DerivedStatFormula } from '../types/world.types';
 import { generateUniqueId } from '../lib/utils/generateId';
@@ -23,79 +22,11 @@ import { storeEvents, StoreEventTypes, type CharacterDeletedEvent, type WorldDel
 import Logger from '@/lib/utils/logger';
 const logger = new Logger('CharacterStore');
 
-// Simplified character types for MVP implementation
-export interface CharacterAttribute {
-  id: EntityID;
-  characterId: EntityID;
-  worldAttributeId?: EntityID; // Reference to world attribute for safer matching
-  name: string;
-  baseValue: number;
-  modifiedValue: number;
-  category?: string;
-}
-
-export interface CharacterSkill {
-  id: EntityID;
-  characterId: EntityID;
-  worldSkillId?: EntityID; // Reference to world skill for safer matching
-  name: string;
-  level: number;
-  category?: string;
-}
-
-// Note: DerivedStat is imported from character.types.ts
-
-interface CharacterBackground {
-  history: string;
-  personality: string;
-  goals: string[];
-  fears: string[];
-  physicalDescription?: string;
-  relationships: unknown[];
-  isKnownFigure?: boolean;
-  knownFigureType?:
-    | 'historical'
-    | 'fictional'
-    | 'celebrity'
-    | 'mythological'
-    | 'other';
-}
-
-interface CharacterStatus {
-  health: number;
-  maxHealth: number;
-  conditions: string[];
-  location?: string;
-}
-
-export interface Character {
-  id: EntityID;
-  name: string;
-  description: string;
-  worldId: EntityID;
-  level: number;
-  attributes: CharacterAttribute[];
-  skills: CharacterSkill[];
-  derivedStats: DerivedStat[];
-  background: CharacterBackground;
-  isPlayer: boolean;
-  status: CharacterStatus;
-  inventory: {
-    characterId: EntityID;
-    items: InventoryItem[];
-    capacity: number;
-    categories: InventoryCategory[];
-    itemOrder: EntityID[];
-  };
-  portrait?: {
-    type: 'ai-generated' | 'placeholder';
-    url: string | null;
-    generatedAt?: string;
-    prompt?: string;
-  };
-  createdAt: string;
-  updatedAt: string;
-}
+// Character shapes live in characterStore.types.ts so lib code can import
+// them without pulling in the store module. Re-exported here to keep the
+// existing import surface working.
+import type { Character, CharacterAttribute, CharacterSkill } from './characterStore.types';
+export type { Character, CharacterAttribute, CharacterSkill } from './characterStore.types';
 
 const addCharacterToRoster = (
   rosters: Record<EntityID, EntityID[]>,
